data_cleaned
Hey this is what i have so far I'm checking out for the night. The student.ipynb file on this branch is updated and if just contains a info from a few data frames that i thought might be valuable as well as a new note file explaining which columns came from which csv file, perhaps that will go in the final readme down the line.